### 2021-06-23 #3396 @whedon
[PRE REVIEW]: Leafmap: A Python package for interactive mapping and geospatial analysis with minimal coding in a Jupyter environment
**Submitting author:** @giswqs (<a href="http://orcid.org/0000-0001-5437-4073">Qiusheng Wu</a>)
**Repository:** <a href="https://github.com/giswqs/leafmap" target ="_blank">https://github.com/giswqs/leafmap</a>
**Version:** v0.3.2
