### 2021-06-25 8f78a22
Added colormap notebook #60
