+++ mkdocs.yml
@@ -82,3 +82,4 @@ nav:
           - notebooks/20_planet_imagery.ipynb
           - notebooks/21_ts_inspector.ipynb
           - notebooks/22_time_slider.ipynb
+          - notebooks/23_colormaps.ipynb
