### 2021-06-26 402d29f
Updated colormap example #60
