### 2021-06-26 9e70d71
Updated colormap example #60
