+++ docs/tutorials.md
@@ -26,6 +26,7 @@
 20. Adding Planet global monthly and quarterly mosaic ([notebook](https://leafmap.org/notebooks/20_planet_imagery))
 21. Using timeseries inspector with one click ([notebook](https://leafmap.org/notebooks/21_ts_inspector))
 22. Using time slider for visualizing timeseries images ([notebook](https://leafmap.org/notebooks/22_time_slider))
+23. Creating colormaps with a single line of code ([notebook](https://leafmap.org/notebooks/23_colormaps))
 
 ## Demo
 
