### 2021-06-28 8d1d4ff
Added cog mosaic from file #61
