### 2021-06-28 b5177ac
Added cog mosaic from file #61
