### 2021-06-28 3e1c2e8
Added cog mosaic from file #61
