+++ leafmap/foliumap.py
@@ -698,7 +698,91 @@ class Map(folium.Map):
             folium.GeoJson(
                 data=fc,
                 # style_function=style_function,
-                name="Footprints",
+                name=name + "_footprints",
+            ).add_to(self)
+
+            center = get_center(fc)
+            if verbose:
+                print("The footprint layer has been added.")
+        else:
+            center = get_cog_center(links[0], titiler_endpoint)
+
+        self.set_center(center[0], center[1], zoom=6)
+
+    def add_cog_mosaic_from_file(
+        self,
+        filepath,
+        skip_rows=0,
+        name="Untitled",
+        attribution=".",
+        opacity=1.0,
+        shown=True,
+        titiler_endpoint="https://api.cogeo.xyz/",
+        username="anonymous",
+        overwrite=False,
+        show_footprints=False,
+        verbose=True,
+        **kwargs,
+    ):
+        """Add a virtual mosaic of COGs to the map.
+
+        Args:
+            filepath (str): Local path or HTTP URL to the csv/txt file containing COG URLs.
+            skip_rows (int, optional): The number of rows to skip in the file. Defaults to 0.
+            name (str, optional): The layer name to use for the layer. Defaults to 'Untitled'.
+            attribution (str, optional): The attribution to use. Defaults to '.'.
+            opacity (float, optional): The opacity of the layer. Defaults to 1.
+            shown (bool, optional): A flag indicating whether the layer should be on by default. Defaults to True.
+            titiler_endpoint (str, optional): Titiler endpoint. Defaults to "https://api.cogeo.xyz/".
+            username (str, optional): The username to create mosaic using the titiler endpoint. Defaults to 'anonymous'.
+            overwrite (bool, optional): Whether or not to replace existing layer with the same layer name. Defaults to False.
+            show_footprints (bool, optional): Whether or not to show footprints of COGs. Defaults to False.
+            verbose (bool, optional): Whether or not to print descriptions. Defaults to True.
+        """
+        import urllib
+
+        layername = name.replace(" ", "_")
+
+        links = []
+        if filepath.startswith("http"):
+            data = urllib.request.urlopen(filepath)
+            for line in data:
+                links.append(line.decode("utf-8").strip())
+
+        else:
+            with open(filepath) as f:
+                links = [line.strip() for line in f.readlines()]
+
+        links = links[skip_rows:]
+
+        tile = get_cog_mosaic(
+            links,
+            titiler_endpoint=titiler_endpoint,
+            username=username,
+            layername=layername,
+            overwrite=overwrite,
+            verbose=verbose,
+        )
+        self.add_tile_layer(tile, name, attribution, opacity, shown)
+
+        if show_footprints:
+            if verbose:
+                print(
+                    f"Generating footprints of {len(links)} COGs. This might take a while ..."
+                )
+            coords = []
+            for link in links:
+                coord = get_cog_bounds(link)
+                if coord is not None:
+                    coords.append(coord)
+            fc = coords_to_geojson(coords)
+
+            # style_function = lambda x: {'opacity': 1, 'dashArray': '1', 'fillOpacity': 0, 'weight': 1}
+
+            folium.GeoJson(
+                data=fc,
+                # style_function=style_function,
+                name=name + "_footprints",
             ).add_to(self)
 
             center = get_center(fc)
