### 2021-06-28 595e01a
Changed COG and STAC function names #61
