+++ leafmap/common.py
@@ -761,7 +761,7 @@ def create_code_cell(code="", where="below"):
     )
 
 
-def get_cog_tile(url, titiler_endpoint="https://api.cogeo.xyz/", **kwargs):
+def cog_tile(url, titiler_endpoint="https://api.cogeo.xyz/", **kwargs):
     """Get a tile layer from a Cloud Optimized GeoTIFF (COG).
         Source code adapted from https://developmentseed.org/titiler/examples/Working_with_CloudOptimizedGeoTIFF_simple/
 
@@ -795,7 +795,7 @@ def get_cog_tile(url, titiler_endpoint="https://api.cogeo.xyz/", **kwargs):
     return r["tiles"][0]
 
 
-def get_cog_mosaic(
+def cog_mosaic(
     links,
     titiler_endpoint="https://api.cogeo.xyz/",
     username="anonymous",
@@ -860,7 +860,7 @@ def get_cog_mosaic(
         raise Exception(e)
 
 
-def get_cog_mosaic_from_file(
+def cog_mosaic_from_file(
     filepath,
     skip_rows=0,
     titiler_endpoint="https://api.cogeo.xyz/",
@@ -898,13 +898,13 @@ def get_cog_mosaic_from_file(
 
     links = links[skip_rows:]
     # print(links)
-    mosaic = get_cog_mosaic(
+    mosaic = cog_mosaic(
         links, titiler_endpoint, username, layername, overwrite, verbose, **kwargs
     )
     return mosaic
 
 
-def get_cog_bounds(url, titiler_endpoint="https://api.cogeo.xyz/"):
+def cog_bounds(url, titiler_endpoint="https://api.cogeo.xyz/"):
     """Get the bounding box of a Cloud Optimized GeoTIFF (COG).
 
     Args:
@@ -925,7 +925,7 @@ def get_cog_bounds(url, titiler_endpoint="https://api.cogeo.xyz/"):
     return bounds
 
 
-def get_cog_center(url, titiler_endpoint="https://api.cogeo.xyz/"):
+def cog_center(url, titiler_endpoint="https://api.cogeo.xyz/"):
     """Get the centroid of a Cloud Optimized GeoTIFF (COG).
 
     Args:
@@ -935,12 +935,12 @@ def get_cog_center(url, titiler_endpoint="https://api.cogeo.xyz/"):
     Returns:
         tuple: A tuple representing (longitude, latitude)
     """
-    bounds = get_cog_bounds(url, titiler_endpoint)
+    bounds = cog_bounds(url, titiler_endpoint)
     center = ((bounds[0] + bounds[2]) / 2, (bounds[1] + bounds[3]) / 2)  # (lat, lon)
     return center
 
 
-def get_cog_bands(url, titiler_endpoint="https://api.cogeo.xyz/"):
+def cog_bands(url, titiler_endpoint="https://api.cogeo.xyz/"):
     """Get band names of a Cloud Optimized GeoTIFF (COG).
 
     Args:
@@ -963,7 +963,7 @@ def get_cog_bands(url, titiler_endpoint="https://api.cogeo.xyz/"):
     return bands
 
 
-def get_stac_tile(url, bands=None, titiler_endpoint="https://api.cogeo.xyz/", **kwargs):
+def stac_tile(url, bands=None, titiler_endpoint="https://api.cogeo.xyz/", **kwargs):
     """Get a tile layer from a single SpatialTemporal Asset Catalog (STAC) item.
 
     Args:
@@ -991,7 +991,7 @@ def get_stac_tile(url, bands=None, titiler_endpoint="https://api.cogeo.xyz/", **
     if "maxzoom" in kwargs.keys():
         params["maxzoom"] = kwargs["maxzoom"]
 
[…]
