+++ leafmap/leafmap.py
@@ -710,8 +710,8 @@ class Map(ipyleaflet.Map):
             shown (bool, optional): A flag indicating whether the layer should be on by default. Defaults to True.
             titiler_endpoint (str, optional): Titiler endpoint. Defaults to "https://api.cogeo.xyz/".
         """
-        tile_url = get_cog_tile(url, titiler_endpoint, **kwargs)
-        center = get_cog_center(url, titiler_endpoint)  # (lon, lat)
+        tile_url = cog_tile(url, titiler_endpoint, **kwargs)
+        center = cog_center(url, titiler_endpoint)  # (lon, lat)
         self.add_tile_layer(tile_url, name, attribution, opacity, shown)
         self.set_center(lon=center[0], lat=center[1], zoom=10)
 
@@ -744,7 +744,7 @@ class Map(ipyleaflet.Map):
             verbose (bool, optional): Whether or not to print descriptions. Defaults to True.
         """
         layername = name.replace(" ", "_")
-        tile = get_cog_mosaic(
+        tile = cog_mosaic(
             links,
             titiler_endpoint=titiler_endpoint,
             username=username,
@@ -761,7 +761,7 @@ class Map(ipyleaflet.Map):
                 )
             coords = []
             for link in links:
-                coord = get_cog_bounds(link)
+                coord = cog_bounds(link)
                 if coord is not None:
                     coords.append(coord)
             fc = coords_to_geojson(coords)
@@ -782,7 +782,7 @@ class Map(ipyleaflet.Map):
             if verbose:
                 print("The footprint layer has been added.")
         else:
-            center = get_cog_center(links[0], titiler_endpoint)
+            center = cog_center(links[0], titiler_endpoint)
 
         self.set_center(center[0], center[1], zoom=6)
 
@@ -832,7 +832,7 @@ class Map(ipyleaflet.Map):
 
         links = links[skip_rows:]
 
-        tile = get_cog_mosaic(
+        tile = cog_mosaic(
             links,
             titiler_endpoint=titiler_endpoint,
             username=username,
@@ -850,7 +850,7 @@ class Map(ipyleaflet.Map):
                 )
             coords = []
             for link in links:
-                coord = get_cog_bounds(link)
+                coord = cog_bounds(link)
                 if coord is not None:
                     coords.append(coord)
             fc = coords_to_geojson(coords)
@@ -871,7 +871,7 @@ class Map(ipyleaflet.Map):
             if verbose:
                 print("The footprint layer has been added.")
         else:
-            center = get_cog_center(links[0], titiler_endpoint)
+            center = cog_center(links[0], titiler_endpoint)
 
         self.set_center(center[0], center[1], zoom=6)
 
@@ -896,8 +896,8 @@ class Map(ipyleaflet.Map):
             shown (bool, optional): A flag indicating whether the layer should be on by default. Defaults to True.
             titiler_endpoint (str, optional): Titiler endpoint. Defaults to "https://api.cogeo.xyz/".
         """
-        tile_url = get_stac_tile(url, bands, titiler_endpoint, **kwargs)
-        center = get_stac_center(url, titiler_endpoint)
+        tile_url = stac_tile(url, bands, titiler_endpoint, **kwargs)
+        center = stac_center(url, titiler_endpoint)
         self.add_tile_layer(tile_url, name, attribution, opacity, shown)
         self.set_center(lon=center[0], lat=center[1], zoom=10)
 
