### 2021-06-28 a394227
Changed COG and STAC function names #61
