+++ leafmap/foliumap.py
@@ -631,8 +631,8 @@ class Map(folium.Map):
             shown (bool, optional): A flag indicating whether the layer should be on by default. Defaults to True.
             titiler_endpoint (str, optional): Titiler endpoint. Defaults to "https://api.cogeo.xyz/".
         """
-        tile_url = get_cog_tile(url, titiler_endpoint, **kwargs)
-        center = get_cog_center(url, titiler_endpoint)  # (lon, lat)
+        tile_url = cog_tile(url, titiler_endpoint, **kwargs)
+        center = cog_center(url, titiler_endpoint)  # (lon, lat)
         self.add_tile_layer(
             url=tile_url,
             name=name,
@@ -671,7 +671,7 @@ class Map(folium.Map):
             verbose (bool, optional): Whether or not to print descriptions. Defaults to True.
         """
         layername = name.replace(" ", "_")
-        tile = get_cog_mosaic(
+        tile = cog_mosaic(
             links,
             titiler_endpoint=titiler_endpoint,
             username=username,
@@ -688,7 +688,7 @@ class Map(folium.Map):
                 )
             coords = []
             for link in links:
-                coord = get_cog_bounds(link)
+                coord = cog_bounds(link)
                 if coord is not None:
                     coords.append(coord)
             fc = coords_to_geojson(coords)
@@ -705,7 +705,7 @@ class Map(folium.Map):
             if verbose:
                 print("The footprint layer has been added.")
         else:
-            center = get_cog_center(links[0], titiler_endpoint)
+            center = cog_center(links[0], titiler_endpoint)
 
         self.set_center(center[0], center[1], zoom=6)
 
@@ -755,7 +755,7 @@ class Map(folium.Map):
 
         links = links[skip_rows:]
 
-        tile = get_cog_mosaic(
+        tile = cog_mosaic(
             links,
             titiler_endpoint=titiler_endpoint,
             username=username,
@@ -772,7 +772,7 @@ class Map(folium.Map):
                 )
             coords = []
             for link in links:
-                coord = get_cog_bounds(link)
+                coord = cog_bounds(link)
                 if coord is not None:
                     coords.append(coord)
             fc = coords_to_geojson(coords)
@@ -789,7 +789,7 @@ class Map(folium.Map):
             if verbose:
                 print("The footprint layer has been added.")
         else:
-            center = get_cog_center(links[0], titiler_endpoint)
+            center = cog_center(links[0], titiler_endpoint)
 
         self.set_center(center[0], center[1], zoom=6)
 
@@ -814,8 +814,8 @@ class Map(folium.Map):
             shown (bool, optional): A flag indicating whether the layer should be on by default. Defaults to True.
             titiler_endpoint (str, optional): Titiler endpoint. Defaults to "https://api.cogeo.xyz/".
         """
-        tile_url = get_stac_tile(url, bands, titiler_endpoint, **kwargs)
-        center = get_stac_center(url, titiler_endpoint)
+        tile_url = stac_tile(url, bands, titiler_endpoint, **kwargs)
+        center = stac_center(url, titiler_endpoint)
         self.add_tile_layer(
             url=tile_url,
             name=name,
