+++ docs/tutorials.md
@@ -27,6 +27,7 @@
 21. Using timeseries inspector with one click ([notebook](https://leafmap.org/notebooks/21_ts_inspector))
 22. Using time slider for visualizing timeseries images ([notebook](https://leafmap.org/notebooks/22_time_slider))
 23. Creating colormaps with a single line of code ([notebook](https://leafmap.org/notebooks/23_colormaps))
+24. Creating heat map from csv ([notebook](https://leafmap.org/notebooks/24_heatmap))
 
 ## Demo
 
