+++ mkdocs.yml
@@ -83,3 +83,4 @@ nav:
           - notebooks/21_ts_inspector.ipynb
           - notebooks/22_time_slider.ipynb
           - notebooks/23_colormaps.ipynb
+          - notebooks/24_heatmap.ipynb
