### 2021-07-01 #3414 @danielskatz
👋 @ajstewartlang - FYI, this is because in https://github.com/openjournals/joss-reviews/issues/3396 you both assigned him and then added him - you only need to do one of these: typically, use assign for the first reviewer, and then add for any subsequent ones.  You can probably use `@whedon remove @TomasBeuzen as reviewer` to remove one of the duplicates here, then regenerate the PDF.  Let me know if you have problems...
