+++ requirements.txt
@@ -1,6 +1,7 @@
 bqplot
 colour
 folium>=0.12.0
+geojson
 googledrivedownloader
 ipyevents
 ipyleaflet
