### 2021-07-07 giswqs/leafmap#76 @martinfleis
leafmap.Map shows only Error displaying widget: model not found
### Environment Information

-   leafmap version: 0.3.2
-   Python version: 3.9.2
-   Operating System: macOS (intel)

### Description

_This is a part of the review for JOSS (https://github.com/openjournals/joss-reviews/issues/3414)_

I installed `leafmap` from conda-forge and tried to follow [Get Started guide](https://leafmap.org/get-started/). However, while using ipyleaflet backend, no map is shown, only a message `Error displaying widget: model not found`.

![Screenshot 2021-07-07 at 16 23 54](https://user-images.githubusercontent.com/36797143/124786577-bb1b6300-df3f-11eb-9c58-324e8acdd335.png)

### What I Did

```
import leafmap

m = leafmap.Map(center=(40, -100), zoom=4)
m
```

Environment specs:

```
# Name                    Version                   Build  Channel
ablog                     0.10.18                  pypi_0    pypi
abseil-cpp                20210324.1           he49afe7_0    conda-forge
affine                    2.3.0                      py_0    conda-forge
alabaster                 0.7.12                     py_0    conda-forge
anyio                     3.2.1            py39h6e9494a_0    conda-forge
appnope                   0.1.2            py39h6e9494a_1    conda-forge
argon2-cffi               20.1.0           py39hcbf5805_2    conda-forge
arrow-cpp                 4.0.0           py39h794a569_3_cpu    conda-forge
async_generator           1.10                       py_0    conda-forge
attrs                     20.3.0             pyhd3deb0d_0    conda-forge
aws-c-cal                 0.5.11               hd2e2f4b_0    conda-forge
aws-c-common              0.6.2                h0d85af4_0    conda-forge
aws-c-event-stream        0.2.7               hb9330a7_13    conda-forge
aws-c-io                  0.10.5               h35aa462_0    conda-forge
aws-checksums             0.1.11               h0010a65_7    conda-forge
aws-sdk-cpp               1.8.186              h766a74d_3    conda-forge
babel                     2.9.1              pyh44b312d_0    conda-forge
backcall                  0.2.0              pyh9f0ad1d_0    conda-forge
backports                 1.0                        py_2    conda-forge
backports.functools_lru_cache 1.6.4              pyhd8ed1ab_0    conda-forge
beautifulsoup4            4.9.3              pyhb0f4dca_0    conda-forge
bleach                    3.3.0              pyh44b312d_0    conda-forge
bokeh                     2.3.2            py39h6e9494a_0    conda-forge
boost-cpp                 1.74.0               hbdcdab7_3    conda-forge
bqplot                    0.12.29            pyhd8ed1ab_0    conda-forge
branca                    0.4.2              pyhd8ed1ab_0    conda-forge
brotli                    1.0.9                h0d85af4_5    conda-forge
brotli-bin                1.0.9                h0d85af4_5    conda-forge
brotlipy                  0.7.0           py39hcbf5805_1001    conda-forge
bs4                       4.9.3                         0    conda-forge
bzip2                     1.0.8                hc929b4f_4    conda-forge
c-ares                    1.17.1               h0d85af4_1    conda-forge
ca-certificates           2021.5.30            h033912b_0    conda-forge
cairo                     1.16.0            he43a7df_1008    conda-forge
cartopy                   0.19.0.post1     py39hada9a11_0    conda-forge
certifi                   2021.5.30        py39h6e9494a_0    conda-forge
cffi                      1.14.5           py39hb71fe58_0    conda-forge
cfitsio                   3.470                h4472c5d_7    conda-forge
cftime                    1.5.0            py39hc89836e_0    conda-forge
chardet                   4.0.0            py39h6e9494a_1    conda-forge
click                     7.1.2              pyh9f0ad1d_0    conda-forge
click-completion          0.5.2            py39h6e9494a_2    conda-forge
click-log                 0.3.2              pyh9f0ad1d_0    conda-forge
click-plugins             1.1.1                      py_0    conda-forge
cligj                     0.7.2              pyhd8ed1ab_0    conda-forge
cloudpickle               1.6.0                      py_0    conda-forge
clustergram               0.4.0+3.g65f2b56.dirty           dev_0    <develop>
colorama                  0.4.4              pyh9f0ad1d_0    conda-forge
colorcet                  2.0.6              pyhd8ed1ab_0    conda-forge
colour                    0.1.5                      py_0    conda-forge
conda                     4.10.3           py39h6e9494a_0    conda-forge
conda-package-handling    1.7.3            py39h89e85a6_0    conda-forge
contextily                1.1.0              pyhd8ed1ab_0    conda-forge
cryptography              3.4.7            py39ha2c9959_0    conda-forge
curl                      7.77.0               hb861fe1_0    conda-forge
cycler                    0.10.0                     py_2    conda-forge
cytoolz                   0.11.0           py39hcbf5805_3    conda-forge
dask                      2021.6.2           pyhd8ed1ab_0    conda-forge
dask-core                 2021.6.2           pyhd8ed1ab_0    conda-forge
dask-geopandas            0.1.0a2                  pypi_0    pypi
datashader                0.13.0             pyh6c4a22f_0    conda-forge
datashape                 0.5.4                      py_1    conda-forge
dbus                      1.13.6               ha13b53f_2    conda-forge
debugpy                   1.3.0            py39h9fcab8e_0    conda-forge
decorator                 4.4.2                      py_0    conda-forge
defusedxml                0.7.1              pyhd8ed1ab_0    conda-forge
deprecated                1.2.12             pyh44b312d_0    conda-forge
deprecation               2.1.0              pyh9f0ad1d_0    conda-forge
descartes                 1.1.0                      py_4    conda-forge
distributed               2021.6.2         py39h6e9494a_0    conda-forge
docutils                  0.16             py39h6e9494a_3    conda-forge
entrypoints               0.3             py39hde42818_1002    conda-forge
et_xmlfile                1.0.1                   py_1001    conda-forge
expat                     2.4.1                he49afe7_0    conda-forge
feedgen                   0.9.0                    pypi_0    pypi
fiona                     1.8.18           py39h673eaac_1    conda-forge
folium                    0.12.0             pyhd8ed1ab_1    conda-forge
fontconfig                2.13.1            h10f422b_1005    conda-forge
freetype                  2.10.4               h4cff582_1    conda-forge
freexl                    1.0.6                h0d85af4_0    conda-forge
fsspec                    2021.6.1           pyhd8ed1ab_0    conda-forge
gdal                      3.2.1            py39h0530131_7    conda-forge
geckodriver               0.29.0               h970cd5f_0    conda-forge
geographiclib             1.52               pyhd8ed1ab_0    conda-forge
geopandas                 0.9.0              pyhd8ed1ab_1    conda-forge
geopandas-base            0.9.0              pyhd8ed1ab_1    conda-forge
geoplot                   0.4.3              pyhd8ed1ab_0    conda-forge
geopy                     2.1.0              pyhd3deb0d_0    conda-forge
geos                      3.9.1                he49afe7_2    conda-forge
geotiff                   1.6.0                h1925f3f_4    conda-forge
geoviews                  1.9.1              pyhd8ed1ab_0    conda-forge
geoviews-core             1.9.1              pyha770c72_0    conda-forge
gettext                   0.19.8.1          h7937167_1005    conda-forge
gflags                    2.2.2             hb1e8313_1004    conda-forge
giflib                    5.2.1                hbcb3906_2    conda-forge
git                       2.30.2          pl5320h2551975_1    conda-forge
gitdb                     4.0.7              pyhd3eb1b0_0
gitpython                 3.1.18             pyhd8ed1ab_0    conda-forge
glib                      2.68.3               he49afe7_0    conda-forge
glib-tools                2.68.3               he49afe7_0    conda-forge
glog                      0.5.0                h25b26a9_0    conda-forge
googledrivedownloader     0.4                pyhd3deb0d_1    conda-forge
greenlet                  1.1.0            py39h9fcab8e_0    conda-forge
grpc-cpp                  1.37.1               h25f885f_2    conda-forge
hdf4                      4.2.15               hefd3b78_3    conda-forge
hdf5                      1.10.6          nompi_hc5d9132_1114    conda-forge
heapdict                  1.0.1                      py_0    conda-forge
here-map-widget-for-jupyter 1.1.1              pyhd8ed1ab_0    conda-forge
holoviews                 1.14.4             pyhd8ed1ab_0    conda-forge
html2text                 2020.1.16                  py_0    conda-forge
hvplot                    0.7.2              pyh6c4a22f_0    conda-forge
icu                       68.1                 h74dc148_0    conda-forge
idna                      2.10               pyh9f0ad1d_0    conda-forge
imagesize                 1.2.0                      py_0    conda-forge
importlib-metadata        4.6.1            py39h6e9494a_0    conda-forge
importlib_metadata        4.6.1                hd8ed1ab_0    conda-forge
invoke                    1.5.0                    pypi_0    pypi
ipyevents                 0.8.2              pyhd8ed1ab_0    conda-forge
ipyfilechooser            0.4.4              pyhd8ed1ab_0    conda-forge
ipykernel                 6.0.1            py39h71a6800_0    conda-forge
ipyleaflet                0.14.0             pyhd8ed1ab_1    conda-forge
ipython                   7.25.0           py39h71a6800_1    conda-forge
ipython_genutils          0.2.0                      py_1    conda-forge
ipytree                   0.2.1              pyhd8ed1ab_0    conda-forge
ipywidgets                7.6.3              pyhd3deb0d_0    conda-forge
jdcal                     1.4.1                      py_0    conda-forge
jedi                      0.18.0           py39h6e9494a_2    conda-forge
jinja2                    3.0.1              pyhd8ed1ab_0    conda-forge
joblib                    1.0.1              pyhd8ed1ab_0    conda-forge
jpeg                      9d                   hbcb3906_0    conda-forge
json-c                    0.15                 hcb556a6_0    conda-forge
json5                     0.9.5              pyh9f0ad1d_0    conda-forge
jsonschema                3.2.0              pyhd8ed1ab_3    conda-forge
jupyter                   1.0.0            py39h6e9494a_6    conda-forge
jupyter-cache             0.4.2              pyhd8ed1ab_0    conda-forge
jupyter-packaging         0.10.4             pyhd8ed1ab_0    conda-forge
jupyter-server-mathjax    0.2.3              pyhd8ed1ab_0    conda-forge
jupyter-sphinx            0.3.2                      py_0    conda-forge
jupyter_client            6.1.12             pyhd8ed1ab_0    conda-forge
jupyter_console           6.4.0              pyhd8ed1ab_0    conda-forge
jupyter_core              4.7.1            py39h6e9494a_0    conda-forge
jupyter_server            1.9.0              pyhd8ed1ab_0    conda-forge
jupyterlab                3.0.16             pyhd8ed1ab_0    conda-forge
jupyterlab_pygments       0.1.2              pyh9f0ad1d_0    conda-forge
jupyterlab_server         2.6.0              pyhd8ed1ab_0    conda-forge
jupyterlab_widgets        1.0.0              pyhd8ed1ab_1    conda-forge
kealib                    1.4.14               h31dd65d_2    conda-forge
kiwisolver                1.3.1            py39hf018cea_1    conda-forge
krb5                      1.19.1               hcfbf3a7_0    conda-forge
lcms2                     2.12                 h577c468_0    conda-forge
leafmap                   0.3.2              pyhd8ed1ab_0    conda-forge
libarchive                3.5.1                h0a5793d_1    conda-forge
libblas                   3.9.0                9_openblas    conda-forge
libbrotlicommon           1.0.9                h0d85af4_5    conda-forge
libbrotlidec              1.0.9                h0d85af4_5    conda-forge
libbrotlienc              1.0.9                h0d85af4_5    conda-forge
libcblas                  3.9.0                9_openblas    conda-forge
libclang                  11.1.0          default_he082bbe_1    conda-forge
libcurl                   7.77.0               hf45b732_0    conda-forge
libcxx                    11.1.0               habf9029_0    conda-forge
libdap4                   3.20.6               h3e144a0_2    conda-forge
libedit                   3.1.20191231         hed1e85f_2    conda-forge
libev                     4.33                 haf1e3a3_1    conda-forge
libevent                  2.1.10               hddc9c9b_3    conda-forge
libffi                    3.3                  h046ec9c_2    conda-forge
libgdal                   3.2.1                h62a8350_7    conda-forge
libgfortran               5.0.0           9_3_0_h6c81a4c_22    conda-forge
libgfortran5              9.3.0               h6c81a4c_22    conda-forge
libglib                   2.68.3               hd556434_0    conda-forge
libiconv                  1.16                 haf1e3a3_0    conda-forge
libkml                    1.3.0             h8fd9edb_1013    conda-forge
liblapack                 3.9.0                9_openblas    conda-forge
libllvm10                 10.0.1               h009f743_3    conda-forge
libllvm11                 11.1.0               hd011deb_2    conda-forge
libnetcdf                 4.7.4           nompi_h9d8a93f_107    conda-forge
libnghttp2                1.43.0               h07e645a_0    conda-forge
libopenblas               0.3.15          openmp_h5e1b9a4_1    conda-forge
libpng                    1.6.37               hb0a8c7a_2    conda-forge
libpq                     13.3                 hea3049e_0    conda-forge
libprotobuf               3.16.0               hcf210ce_0    conda-forge
libpysal                  4.5.1              pyhd8ed1ab_0    conda-forge
librttopo                 1.1.0                h5413771_6    conda-forge
libsodium                 1.0.18               hbcb3906_1    conda-forge
libsolv                   0.7.19               hcf210ce_3    conda-forge
libspatialindex           1.9.3                h4a8c4bd_3    conda-forge
libspatialite             5.0.1                h395434c_3    conda-forge
libssh2                   1.9.0                h52ee1ee_6    conda-forge
libthrift                 0.14.2               hab56fdc_1    conda-forge
libtiff                   4.2.0                h46d1c8c_2    conda-forge
libutf8proc               2.6.1                h35c211d_0    conda-forge
libwebp-base              1.2.0                h0d85af4_2    conda-forge
libxcb                    1.13              h35c211d_1003    conda-forge
libxml2                   2.9.12               h93ec3fd_0    conda-forge
llvm-openmp               11.1.0               hda6cdc1_1    conda-forge
llvmlite                  0.36.0           py39h798a4f4_0    conda-forge
locket                    0.2.0                      py_2    conda-forge
lxml                      4.6.3                    pypi_0    pypi
lz4-c                     1.9.3                h046ec9c_0    conda-forge
lzo                       2.10              haf1e3a3_1000    conda-forge
mamba                     0.14.1           py39hb671511_0    conda-forge
mapclassify               2.4.2              pyhd8ed1ab_0    conda-forge
markdown                  3.3.4              pyhd8ed1ab_0    conda-forge
markdown-it-py            0.6.2              pyhd8ed1ab_0    conda-forge
markupsafe                2.0.1            py39h89e85a6_0    conda-forge
matplotlib                3.4.2            py39h6e9494a_0    conda-forge
matplotlib-base           3.4.2            py39hb07454d_0    conda-forge
matplotlib-inline         0.1.2              pyhd8ed1ab_2    conda-forge
mdit-py-plugins           0.2.6              pyhd8ed1ab_0    conda-forge
mercantile                1.2.1              pyhd8ed1ab_0    conda-forge
mistune                   0.8.4           py39h89e85a6_1004    conda-forge
momepy                    0.4.4              pyhd8ed1ab_0    conda-forge
msgpack-python            1.0.2            py39hf018cea_1    conda-forge
multipledispatch          0.6.0                      py_0    conda-forge
munch                     2.5.0                      py_0    conda-forge
mysql-common              8.0.25               h694c41f_0    conda-forge
mysql-libs                8.0.25               h54f5a68_0    conda-forge
myst-nb                   0.12.3             pyhd8ed1ab_0    conda-forge
myst-parser               0.13.7             pyhd8ed1ab_0    conda-forge
nbclassic                 0.3.1              pyhd8ed1ab_1    conda-forge
nbclient                  0.5.3              pyhd8ed1ab_0    conda-forge
nbconvert                 5.6.0                      py_0    conda-forge
nbdime                    3.1.0              pyhd8ed1ab_0    conda-forge
nbformat                  5.1.3              pyhd8ed1ab_0    conda-forge
nbsphinx                  0.8.6              pyhd8ed1ab_1    conda-forge
ncurses                   6.2                  h2e338ed_4    conda-forge
nest-asyncio              1.5.1              pyhd8ed1ab_0    conda-forge
netcdf4                   1.5.6           nompi_py39h353b61e_102    conda-forge
networkx                  2.5.1              pyhd8ed1ab_0    conda-forge
notebook                  6.4.0              pyha770c72_0    conda-forge
nspr                      4.30                 hcd9eead_0    conda-forge
nss                       3.67                 h31e2bf1_0    conda-forge
numba                     0.53.1           py39h32e38f5_1    conda-forge
numpy                     1.21.0           py39h7eed0ac_0    conda-forge
numpydoc                  1.1.0                      py_1    conda-forge
olefile                   0.46               pyh9f0ad1d_1    conda-forge
openjpeg                  2.4.0                h6e7aa92_1    conda-forge
openpyxl                  3.0.7              pyhd8ed1ab_0    conda-forge
openssl                   1.1.1k               h0d85af4_0    conda-forge
orc                       1.6.7                h590cbdb_2    conda-forge
osmnx                     1.1.1              pyhd8ed1ab_0    conda-forge
packaging                 21.0               pyhd8ed1ab_0    conda-forge
palettable                3.3.0                      py_0    conda-forge
pandas                    1.3.0            py39h4d6be9b_0    conda-forge
pandoc                    2.14.0.3             h0d85af4_0    conda-forge
pandocfilters             1.4.2                      py_1    conda-forge
panel                     0.11.3             pyhd8ed1ab_0    conda-forge
param                     1.11.1             pyh6c4a22f_0    conda-forge
parquet-cpp               1.5.1                         1    conda-forge
parso                     0.8.2              pyhd8ed1ab_0    conda-forge
partd                     1.2.0              pyhd8ed1ab_0    conda-forge
patsy                     0.5.1                      py_0    conda-forge
pcre                      8.45                 he49afe7_0    conda-forge
perl                      5.32.1          0_h0d85af4_perl5    conda-forge
pexpect                   4.8.0              pyh9f0ad1d_2    conda-forge
pickleshare               0.7.5           py39hde42818_1002    conda-forge
pillow                    8.2.0            py39h5fdd921_1    conda-forge
pip                       21.1.3             pyhd8ed1ab_0    conda-forge
pixman                    0.40.0               hbcb3906_0    conda-forge
poppler                   0.89.0               hd735947_5    conda-forge
poppler-data              0.4.10                        0    conda-forge
postgresql                13.3                 he8fe76e_0    conda-forge
proj                      7.2.0                h78d1473_2    conda-forge
prometheus_client         0.11.0             pyhd8ed1ab_0    conda-forge
[…]
