### 2021-07-07 giswqs/leafmap#76 @giswqs
This a well-known issue related to ipyleaflet.

- https://github.com/jupyter-widgets/ipyleaflet/issues/739
- https://github.com/jupyter-widgets/ipyleaflet/issues/418
- https://github.com/jupyter-widgets/ipyleaflet/issues/504)

Could you try creating a fresh conda env to install leafmap? See https://leafmap.org/installation/#install-from-conda-forge

```
    conda create -n geo python=3.8
    conda activate geo
    conda install geopandas
    conda install mamba -c conda-forge
    mamba install leafmap xarray_leaflet -c conda-forge
```

Alternatively, you can run leafmap directly using binder:

- https://gishub.org/leafmap-pangeo
- https://gishub.org/leafmap-binder
