### 2021-07-07 giswqs/leafmap#76 @martinfleis
Thank you. Running `jupyter labextension install jupyter-leaflet` fixed the issue.

In any case, I would suggest adding a troubleshooting section to the installation docs mentioning this, since it is _a well-known issue_.
