### 2021-07-07 giswqs/leafmap#76 @martinfleis
Ah, now I see that it is actually mentioned in FAQ (https://leafmap.org/faq/#why-the-interactive-map-does-not-show-up). Might be better to expand (include the actual error message so it is searchable) on it and link it from installation docs.
