### 2021-07-07 giswqs/leafmap#76 @giswqs
Good suggestion. I have added a troubleshooting section to the installation and faq pages, which include the actual error message and it is searchable.

- https://leafmap.org/installation/#troubleshooting
- https://leafmap.org/faq/#why-the-interactive-map-does-not-show-up
