### 2021-07-08 giswqs/leafmap#76 @martinfleis
Perfect! Thanks a lot.
