+++ docs/changelog.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## v0.3.3 - July 8, 2021
+
+**New Features**:
+
+-   Added troubleshooting section [#76](https://github.com/giswqs/leafmap/issues/76)
+-   Added pandas_to_geojson function [#75](https://github.com/giswqs/leafmap/issues/75)
+-   Added creating heat map from csv [#64](https://github.com/giswqs/leafmap/issues/64)
+-   Added cog mosaic from file [#61](https://github.com/giswqs/leafmap/issues/61)
+-   Added colormap notebook [#60](https://github.com/giswqs/leafmap/issues/60)
+
+**Improvement**
+
+-   Changed COG and STAC function names [#61](https://github.com/giswqs/leafmap/issues/61)
+-   Updated colormap example [#60](https://github.com/giswqs/leafmap/issues/60)
+
 ## v0.3.2 - June 22, 2021
 
 **New Features**:
