### 2021-07-09 #3414 @martinfleis
Dear @giswqs and @ajstewartlang,

I have finished the review of the package and its paper. Even though the package is an advanced piece of software helpful on many occasions, some issues need to be resolved before publication in JOSS. Please find a detailed explanation below.

### Review

`leafmap` will undoubtedly find its audience. It does what it promises to do and does it well. I have only a few remarks, mostly minor, but let me start with the only major issue.

The software, in its current state, can be considered untested. The repository contains a few test files, but those do not sufficiently check the functionality. As reported by `pytest-cov`, overall coverage is only 12% (see the full report below). I would argue that even those 12% do not reflect the actual state as most of the tests are essentially checking *if it fails or not.* Test suite needs to check whether the output of each class, method and function returns what is expected of it. The current test suite does not do that. I understand that it may not be as straightforward in interactive mapping, but there are ways based on string (HTML) outputs - look at `folium`, for example.

```bash
Name                   Stmts   Miss  Cover
------------------------------------------
leafmap/__init__.py       16      3    81%
leafmap/basemaps.py       28      4    86%
leafmap/colormaps.py      74     74     0%
leafmap/common.py       1038    839    19%
leafmap/foliumap.py      470    470     0%
leafmap/heremap.py       153    153     0%
leafmap/leafmap.py       932    809    13%
leafmap/legends.py        27     24    11%
leafmap/osm.py            88     50    43%
leafmap/toolbar.py       549    549     0%
------------------------------------------
TOTAL                   3375   2975    12%
```

Now on to minor issues (in no particular order):

- XYZ tiles require attribution. Without requiring one, `leafmap` may violate terms of use.
- I would welcome some explanation of different backends in the introduction, not only in the FAQ. In the Get Started notebook, there is no explanation, leaving the newcomer confused. It should also be mentioned that backends do not offer equal functionality.
- Methods supported by one backend and not the other should raise an informative warning, as `NotImplementedError` with troubleshooting (use the other backend), instead of an `AttributeError`.

```bash
AttributeError: module 'leafmap.foliumap' has no attribute 'linked_maps'`
```

- Most of the notebooks in the Tutorials section of the documentation would benefit from some descriptive text accompanying the code. Unfortunately, many of them do not contain any, and not all of them are self-explanatory.
- Just for a reference, one of my earlier issues has already been resolved ().

The mental model `leafmap` use is very different from the PyData stack I am mostly used to. It tries to solve *everything,* which is even taken into extremes with functions like `csv_to_pandas` which is just `pd.read_csv(in_csv, **kwargs)` or `read_postgis` which is a wrapper of `gpd.read_postgis(sql, con, geom_col, crs, **kwargs)`. It is not a problem that they exist; I am just struggling to figure out whether they should be a part of the public API as they are now or whether they are considered private. The same applies to a series of functions I would not expect in the API of a package like `leafmap`, supporting the dev-level installation of packages (`update_package`, `install_from_github`...). Again, should it be part of a user-facing API? I am not asking for these to be changed; I just wanted to point out that it is not a very common practice.

### Paper

A few notes regarding the paper:

- The usage of bullet points with hyperlinks and without much description is a rather strange choice. I would recommend reformulating those sections. The paper needs to be readable without the necessity to click on any link in there.
- The package itself is only briefly explained. A paragraph with a more detailed description outlining different modules, backends and internal logic of the library would help.
- folium should be cited
- ipyleaflet should be cited
- ipywidgets should be cited
