### 2021-07-09 #3414 @giswqs
@martinfleis Thank you very much for the thorough review and constructive comments. I will revise the repo and paper as suggested over the next few days.

Regarding your major comment on test coverage, I will look into the `folium` package to see how to increase the test coverage in interactive mapping. As you already pointed out, test coverage in interactive mapping is not very straightforward. I followed the [ipyleaflet](https://github.com/jupyter-widgets/ipyleaflet) package for testing coverage in interactive mapping. The ipyleaflet repo itself does not contain the `tests` folder. All the tests are actually performed when the documentation website is being built and deployed to https://ipyleaflet.readthedocs.io, which is similar to what I did for leafmap. All the notebook examples demonstrating key functions of leamap are executed and tested using GitHub workflow ([docs.yml](https://github.com/giswqs/leafmap/blob/master/.github/workflows/docs.yml)) when deployed to GitHub Pages, e.g., https://leafmap.org/notebooks/01_leafmap_intro. Users can see the interactive map. The downside is that pytest coverage can't capture these automated tests passed by GitHub workflow.
