### 2021-07-09 #3414 @ajstewartlang
> 👋 @ajstewartlang - FYI, this is because in #3396 you both assigned him and then added him - you only need to do one of these: typically, use assign for the first reviewer, and then add for any subsequent ones. You can probably use `@whedon remove @TomasBeuzen as reviewer` to remove one of the duplicates here, then regenerate the PDF. Let me know if you have problems...

Thanks @danielskatz - yes, looks like I used `assign` twice rather than `add` the second time - I will remove the duplicate and regenerate the PDF....
