### 2021-07-09 #3414 @martinfleis
> All the tests are actually performed when the documentation website is being built

That is interesting, I did not realise that. But that again only checks if the map is generated, not what it shows, right? The point of CI is to ensure the correctness of the output.

A lot of your methods can be tested independently even before they're added to the map itself (e.g. everything in `colormaps.py`), which can make it a bit easier. I've been implementing folium-based plotting to geopandas recently so you may even check how we deal with tests there (we essentially get the HTML string and check if expected strings are part of it) - https://github.com/geopandas/geopandas/pull/1953.
