### 2021-07-09 #3414 @giswqs
@martinfleis It can check if the map is generated as well as functions within the `Map` class. Take a look at this notebook https://leafmap.org/notebooks/10_add_vector as an example. It tests functions such as `add_shp`, `add_geojson`, `add_kml`.  One can verify whether data are being loaded correctly on the map. Leafmap also has a [`add_gdf`](https://leafmap.org/foliumap/#leafmap.foliumap.Map.add_gdf) function for loading GeoPandas GeoDataFrame. I will add it to the notebook example as well.

Thank you for providing the reference to your geopandas folium example. I will learn from it to increase the test coverage.
