### 2021-07-09 #3414 @martinfleis
I know that the example uses those methods and they're executed during the docs creation but there's no automatic check if the result is correct, right? Just that it didn't fail. You can always check manually but that is not what CI is about. Correct me if I misunderstood what is going on there.
