### 2021-07-09 #3414 @giswqs
@martinfleis Yes, I totally agree with you. There is no automatic check if the result is expected during the docs creation. I will learn from your [geopands/test_explore.py](https://github.com/geopandas/geopandas/blob/1e1a861b95ef6f0fc9f841bf6f133db0eb9abdc6/geopandas/tests/test_explore.py) and try my best to increase the test coverage. Getting the HTML string and checking if expected strings are part of it sounds great. One thing to keep in mind is that checking parting of the HTML string does not guarantee that the map will be created successfully or the data will be loaded properly. Leafmap has multiple plotting backends (i.e., ipyleaflet, folium, heremap). The HTML string of the map created using different backends (and data loaded using the same function, e.g., add_geojson, add_shp) might have some common strings in the HTML, but there are not interchangeable. I will use CI as the automatic check and use notebooks during docs creation as a supplemental check to make sure the generated results are as expected.
