### 2021-07-09 #3414 @giswqs
Just to add a reference to the point I raised earlier - checking the HTML string does not guarantee the map/data will be rendered successfully. Here is a relevant ipyleaflet issue (https://github.com/jupyter-widgets/ipyleaflet/issues/841) I opened last month. The map can be exported to HTML without any problem (i.e., you can get the HTML string), but the map still fails to render due to other issues. Therefore, CI would not be able to capture this issue. In this case, using both CI and notebooks executed during docs creation would be very helpful, even though we have to manually check the map/data shown on notebooks.
