+++ .github/workflows/docs.yml
@@ -14,9 +14,9 @@ jobs:
             - name: Install dependencies
               run: |
                   python -m pip install --upgrade pip
-                  pip install --user --no-cache-dir Cython
-                  pip install --user -r requirements.txt
-                  pip install --user .
+                  pip install --no-cache-dir Cython
+                  pip install -r requirements.txt
+                  pip install .
             - name: PKG-TEST
               run: |
                   pip install geopandas osmnx
