### 2021-07-11 giswqs/leafmap#83 @giswqs
Address JOSS reviewer comments
This issue is for addressing the JOSS reviewer comments:

- https://github.com/openjournals/joss-reviews/issues/3414#issuecomment-877210440
- https://github.com/openjournals/joss-reviews/issues/3414#issuecomment-879660026
