### 2021-07-12 7a7c2d9
Added citations #83
