### 2021-07-12 a55f2f1
Added citations #83
