+++ paper/paper.bib
@@ -61,3 +61,46 @@
   url    = "https://zenodo.org/record/4569086",
   doi    = "10.5281/zenodo.4569086"
 }
+
+@MISC{Filipe2021,
+  title  = "{python-visualization/folium v0.12.1}",
+  author = "{Filipe} and Journois, Martin and {Frank} and Story, Rob and
+            Gardiner, James and Rump, Halfdan and Bird, Andrew and Lima,
+            Antonio and Cano, Joshua and {dbf} and Leonel, Juliana and Baker,
+            Jason and Sampson, Tim and Reades, Jon and Welsh, Ben and Kong,
+            Qingkai and Komarov, Oleg and Crosby, Alex and Harris, George and
+            Dumas, Raphael and Krief, David and Kato, Daisuke and
+            {penguindustin} and Wilson, Nat and Nogueira, Tales Paiva and
+            {kenmatsu} and Furtado, Leonardo and Patil, Anand and Duke, Justin
+            and Signell, Rich",
+  month  =  "18~" # jan,
+  year   =  2021,
+  url    = "https://zenodo.org/record/4447642",
+  doi    = "10.5281/zenodo.4447642"
+}
+
+@MISC{Grout2021,
+  title       = "{ipywidgets: Interactive HTML Widgets}",
+  author      = "Grout, Jason and {et al.}",
+  institution = "Github",
+  year        =  2021,
+  url         = "https://github.com/jupyter-widgets/ipywidgets"
+}
+
+@MISC{Renou2021,
+  title       = "{ipyleaflet: A Jupyter / Leaflet bridge enabling interactive
+                 maps in the Jupyter notebook}",
+  author      = "Renou, Martin and Corlay, Sylvain and Brochart, David and {et
+                 al.}",
+  institution = "Github",
+  year        =  2021,
+  url         = "https://github.com/jupyter-widgets/ipyleaflet"
+}
+
+@MISC{Kharude2021,
+  title       = "{here-map-widget-for-jupyter}",
+  author      = "Kharude, Sachin and Steenbergen, Thomas",
+  institution = "Github",
+  year        =  2021,
+  url         = "https://github.com/heremaps/here-map-widget-for-jupyter"
+}
