### 2021-07-12 438d012
Made XYZ tiles attribution required #83
