### 2021-07-12 2e8ebf4
Made XYZ tiles attribution required #83
