### 2021-07-12 3c8872d
Made XYZ tiles attribution required #83
