### 2021-07-13 65f476b
Added map title function #84
