### 2021-07-13 59e19ec
Added map title function #84
