+++ mkdocs.yml
@@ -84,3 +84,4 @@ nav:
           - notebooks/22_time_slider.ipynb
           - notebooks/23_colormaps.ipynb
           - notebooks/24_heatmap.ipynb
+          - notebooks/25_map_title.ipynb
