### 2021-07-13 cbf7b3a
Changed some functions to be private #83
