### 2021-07-13 6d11d3e
Changed some functions to be private #83
