+++ leafmap/__init__.py
@@ -7,7 +7,7 @@ __version__ = "0.3.3"
 import os
 
 
-def in_colab_shell():
+def __in_colab_shell():
     """Tests if the code is being executed within Google Colab."""
     import sys
 
@@ -17,7 +17,7 @@ def in_colab_shell():
         return False
 
 
-def use_folium():
+def __use_folium():
     """Whether to use the folium or ipyleaflet plotting backend."""
     if os.environ.get("USE_FOLIUM") is not None:
         return True
@@ -25,7 +25,7 @@ def use_folium():
         return False
 
 
-if in_colab_shell() or use_folium():
+if __in_colab_shell() or __use_folium():
     from .foliumap import *
 else:
     from .leafmap import *
