### 2021-07-13 433534a
Changed some functions to be private #83
