### 2021-07-14 fa25214
Added more info about plotting backends #83
