### 2021-07-14 c940dc8
Added more info about plotting backends #83
