+++ docs/faq.md
@@ -4,11 +4,18 @@
 
 Please go to <https://github.com/giswqs/leafmap/issues>.
 
-## Why does leafmap use two plotting backends: folium and ipyleaflet
+## What's the difference between folium and ipyleaflet
 
-A key difference between [ipyleaflet](https://github.com/jupyter-widgets/ipyleaflet) and [folium](https://github.com/python-visualization/folium) is that ipyleaflet is built upon ipywidgets and allows bidirectional communication between the front-end and the backend enabling the use of the map to capture user input, while folium is meant for displaying static data only ([source](https://blog.jupyter.org/interactive-gis-in-jupyter-with-ipyleaflet-52f9657fa7a)). Note that [Google Colab](https://colab.research.google.com/) currently does not support ipyleaflet ([source](https://github.com/googlecolab/colabtools/issues/60#issuecomment-596225619)). Therefore, if you are using leafmap
-with Google Colab, you should use `import leafmap.foliumap as leafmap`. If you are using leafmap with a local Jupyter notebook server, you can
-use `import leafmap`, which provides more functionalities for capturing user input (e.g., mouse-clicking and moving).
+A key difference between [ipyleaflet](https://github.com/jupyter-widgets/ipyleaflet) and [folium](https://github.com/python-visualization/folium) is that ipyleaflet is built upon ipywidgets and allows bidirectional communication between the front-end and the backend enabling the use of the map to capture user input, while folium is meant for displaying static data only ([source](https://blog.jupyter.org/interactive-gis-in-jupyter-with-ipyleaflet-52f9657fa7a)). Note that [Google Colab](https://colab.research.google.com/) currently does not support ipyleaflet ([source](https://github.com/googlecolab/colabtools/issues/498#issuecomment-695335421)). Therefore, if you are using leafmap
+with Google Colab, `import leafmap` will automatically use the `folium` plotting backend. If you are using leafmap with Jupyter installed locally, `import leafmap` will automatically use the `ipyleaflet', which provides more functionalities for capturing user input (e.g., mouse-clicking and moving).
+
+## How to use a specific plotting backend
+
+`leafmap` has three plotting backends: [folium](https://github.com/python-visualization/folium), [ipyleaflet](https://github.com/jupyter-widgets/ipyleaflet), and [here-map-widget-for-jupyter](https://github.com/heremaps/here-map-widget-for-jupyter). If you are using `leafmap` with Jupyter installed locally, `import leafmap` will automatically use the `ipyleaflet` plotting backend. If you are using `leafmap` with [Google Colab](https://githubtocolab.com/giswqs/leafmap/blob/master/examples/notebooks/01_leafmap_intro.ipynb), `import leafmap` will automatically use the `folium` plotting backend. Note that Google Colab does not yet support `ipyleaflet` ([source](https://github.com/googlecolab/colabtools/issues/498#issuecomment-695335421)). Therefore, you won't be able to access the `leafmap` toolbar in Colab. Note that the backends do not offer equal functionality. Some interactive functionality in `ipyleaflet` might not be available in `folium` or `heremap`. To use a specific plotting backend, use one of the following:
+
+-   `import leafmap.leafmap as leafmap`
+-   `import leafmap.foliumap as leafmap`
+-   `import leafmap.heremap as leafmap`
 
 ## Why the interactive map does not show up
 
