### 2021-07-14 7ecab18
Added more info about plotting backends #83
