### 2021-07-14 #3414 @TomasBeuzen
Hi @giswqs and @ajstewartlang 👋

I have finished the review of the Leafmap package and its paper. The package and its documentation are comprehensive. I feel it's particularly suited to less-technical users, or those just getting started, but that's not to say it's also a useful tool for more seasoned programmers/developers. Visualizing geospatial data can certainly be a daunting task for novice users so I think this is a nice contribution to the open source world, well done @giswqs. I believe there is a bit of room for improvement before publication, and my comments mostly echo what Martin has said before me.

## Software

- As Martin pointed out before me, I was also concerned about the test suite and coverage, but there has been some discussion about that above and I believe @giswqs is working on this. I look forward to seeing developments in this area.
- I found the source code itself to be well written, nicely formatted, and compartmentalised. The package has a lot of great and very useful functionality and I enjoyed working with the API. It certainly tries to do everything - I was surprised to find simple wrapper functions for things like creating DataFrames or GeoDataFrames, but I suppose these functions are well suited to the niche of novice users that are a target of this package. I'm fine with them being part of the user-facing API.
- In the "Get Started" page your write "Note that the backends do not offer equal functionality." - seeing as there may be novice users leveraging this package, perhaps you could offer a bit more advice about which backend to use for different situations?
- I’d also like to see the “Get Started” and “Usage” pages of the docs rendered with outputs (as much as possible without sharing your API key). It's not super useful to see a heading like "Create an interactive map" followed by some Python code without any output.
- It's great to see so many notebook examples (and youtube videos!) so kudos for that, but more text description is required in these notebooks (I know it’s a lot of work but I feel that having text descriptions of what’s going on is invaluable to new users)
- I think it's important to add a link to the code of conduct in the documentation (perhaps in the "contributing" page)

## Paper

Overall I felt the paper could be expanded on a little:

- I’d like a bit more description about the package itself, i.e., it’s composition/modules, and internal workings. Is it just convenience that the package provides? What does it offer non-novice coders? Along the same vein, could you be explicit about what users you're targeting with the package?
- I don't like the use of dot-points and hyperlinks here, I'd prefer more text describing what the package can do.
- I'd like to see a more explicit  “Comparison to other tools” in the paper. You mention some of the advantages of leafmap in the paper, but I’d like reference to specific tools, e.g., what does leafmap offer over other interactive mapping tools such as  plotly + mapbox, or kepler.gl.
- A few references missing: e.g., folium, ipyleaflet
- A figure or two would be nice but that's personal preference and I wouldn't be upset if you didn't include any.
