### 2021-07-14 #3414 @ajstewartlang
Many thanks for your detailed and helpful reviews @martinfleis and @TomasBeuzen. This looks like a really nice package @giswqs and a contribution that will be really useful for the community. The comments by the reviewers look very helpful and hopefully easy to address. If you could let us know when you've been able to incorporate those suggestions that would be great.
