### 2021-07-14 b07f6a8
Added text description to notebooks #83
