### 2021-07-14 cbf506a
Added text description to notebooks #83
