### 2021-07-14 #3414 @giswqs
@TomasBeuzen Thank you for the thorough review and encouraging comments. I will be addressing your comments in this issue https://github.com/giswqs/leafmap/issues/83. Thank you for your understanding of some of the simple wrapper functions (e.g., `csv_to_pdf`, `read_postgis`). Those functions were designed for novice users who might not be familiar with `pandas` or `geopandas`. They can use `leafmap` without having to import `pandas` or `geopandas`. Other less commonly used functions have been changed to private functions as suggested by @martinfleis. See https://github.com/giswqs/leafmap/commit/a28e84a5989344a99ce93f0a978f89c1e557ddf1. I prefer to keep the `leafmap.update_package()` function, which is essentially `pip install git+https://github.com/giswqs/leafmap`. Typing a function with 2-3 characters and then use autocompletion in Jupyter is much more convenient than memorizing and typing (or copying and pasting) a long string with ~50 characters.

@ajstewartlang I expect to finish the revision by early next week. I will give an update here when it is ready for re-review.
