### 2021-07-14 22800c9
Added NotImplementedError for foliumap #83
