### 2021-07-14 cc63f23
Added NotImplementedError for foliumap #83
