### 2021-07-14 7e418d5
Added NotImplementedError for foliumap #83
