+++ leafmap/colormaps.py
@@ -1,4 +1,4 @@
-"""Module for commonly used colormaps and palettes for visualizing Earth Engine data.
+"""Module for commonly used colormaps and palettes for visualizing geospatial data.
 """
 import matplotlib as mpl
 import matplotlib.pyplot as plt
