### 2021-07-14 f3debaa
fix some of the typo in doc
