### 2021-07-15 58927a5
Added text description to notebooks #83
