### 2021-07-15 a168e3f
Fixed typos using codespell #83
