+++ mkdocs.yml
@@ -17,6 +17,10 @@ theme:
         # - navigation.tabs
     custom_dir: "docs/overrides"
 
+# extra_javascript:
+#     - javascripts/embed-amd.js
+#     - javascripts/require.min.js
+
 plugins:
     - search
     - mkdocstrings
