+++ .github/workflows/docs.yml
@@ -17,6 +17,8 @@ jobs:
                   pip install --no-cache-dir Cython
                   pip install -r requirements.txt
                   pip install .
+            - name: Discover typos with codespell
+              run: codespell --skip="*.csv,*.geojson,*.json,*.js" --ignore-words-list="aci,acount,acounts,fallow,hart,hist,nd,ned,ois,wqs"
             - name: PKG-TEST
               run: |
                   pip install geopandas osmnx
