### 2021-07-15 b6d3de7
Added codespell to docs.yml
