+++ .github/workflows/docs.yml
@@ -15,7 +15,7 @@ jobs:
               run: |
                   python -m pip install --upgrade pip
                   pip install --no-cache-dir Cython
-                  pip install -r requirements.txt
+                  pip install codespell -r requirements.txt
                   pip install .
             - name: Discover typos with codespell
               run: codespell --skip="*.csv,*.geojson,*.json,*.js" --ignore-words-list="aci,acount,acounts,fallow,hart,hist,nd,ned,ois,wqs"
