### 2021-07-17 f4484cf
Added text description to notebooks #83
