### 2021-07-17 1ac8315
Added text description to notebooks #83
