### 2021-07-17 3d186e5
Added text description to notebooks #83
