+++ docs/tutorials.md
@@ -19,12 +19,12 @@
 13. Adding a GeoPandas GeoDataFrame to the map with a single line of code ([notebook](https://leafmap.org/notebooks/13_geopandas))
 14. Adding data from a PostGIS database to the map ([notebook](https://leafmap.org/notebooks/14_postgis))
 15. Downloading OpenStreetMap data with a single line of code ([notebook](https://leafmap.org/notebooks/15_openstreetmap))
-16. Use [HERE Map Widget for Jupyter](https://github.com/heremaps/here-map-widget-for-jupyter) as plotting backend ([notebook](https://leafmap.org/notebooks/16_heremap))
+16. Using [HERE Map Widget for Jupyter](https://github.com/heremaps/here-map-widget-for-jupyter) as a plotting backend ([notebook](https://leafmap.org/notebooks/16_heremap))
 17. Adding vector tile layers to the map ([notebook](https://leafmap.org/notebooks/17_vector_tile_layer))
 18. Adding a point layer with popup attributes to the map ([notebook](https://leafmap.org/notebooks/18_point_layer))
 19. Saving maps as a html file ([notebook](https://leafmap.org/notebooks/19_map_to_html))
 20. Adding Planet global monthly and quarterly mosaic ([notebook](https://leafmap.org/notebooks/20_planet_imagery))
-21. Using timeseries inspector with one click ([notebook](https://leafmap.org/notebooks/21_ts_inspector))
+21. Using timeseries inspector with a single click ([notebook](https://leafmap.org/notebooks/21_ts_inspector))
 22. Using time slider for visualizing timeseries images ([notebook](https://leafmap.org/notebooks/22_time_slider))
 23. Creating colormaps with a single line of code ([notebook](https://leafmap.org/notebooks/23_colormaps))
 24. Creating heat map from csv ([notebook](https://leafmap.org/notebooks/24_heatmap))
