### 2021-07-17 da7302a
Fixed LGTM alerts
