### 2021-07-17 58b4cb3
Fixed LGTM alerts
