### 2021-07-17 8818ac9
Fixed LGTM alerts
