+++ leafmap/leafmap.py
@@ -2,7 +2,6 @@
 
 import os
 import ipyleaflet
-from ipyleaflet import Marker, MarkerCluster, TileLayer, WidgetControl, VectorTileLayer
 from IPython.display import display
 from .basemaps import basemap_tiles
 from .common import *
@@ -354,7 +353,7 @@ class Map(ipyleaflet.Map):
             shown (bool, optional): A flag indicating whether the layer should be on by default. Defaults to True.
         """
         try:
-            tile_layer = TileLayer(
+            tile_layer = ipyleaflet.TileLayer(
                 url=url,
                 name=name,
                 attribution=attribution,
@@ -384,7 +383,7 @@ class Map(ipyleaflet.Map):
             vector_tile_layer_styles(dict,optional): Style dict, specific to the vector tile source.
         """
         try:
-            vector_tile_layer = VectorTileLayer(
+            vector_tile_layer = ipyleaflet.VectorTileLayer(
                 url=url,
                 attribution=attribution,
                 vector_tile_layer_styles=vector_tile_layer_styles,
@@ -919,7 +918,7 @@ class Map(ipyleaflet.Map):
         minimap.layout.width = "150px"
         minimap.layout.height = "150px"
         ipyleaflet.link((minimap, "center"), (self, "center"))
-        minimap_control = WidgetControl(widget=minimap, position=position)
+        minimap_control = ipyleaflet.WidgetControl(widget=minimap, position=position)
         self.add_control(minimap_control)
 
     def add_maker_cluster(self, event="click", add_marker=True):
@@ -934,7 +933,7 @@ class Map(ipyleaflet.Map):
         """
         coordinates = []
         markers = []
-        marker_cluster = MarkerCluster(name="Marker Cluster")
+        marker_cluster = ipyleaflet.MarkerCluster(name="Marker Cluster")
         self.last_click = []
         self.all_clicks = []
         if add_marker:
@@ -948,7 +947,7 @@ class Map(ipyleaflet.Map):
                 self.last_click = latlon
                 self.all_clicks = coordinates
                 if add_marker:
-                    markers.append(Marker(location=latlon))
+                    markers.append(ipyleaflet.Marker(location=latlon))
                     marker_cluster.markers = markers
             elif kwargs.get("type") == "mousemove":
                 pass
@@ -994,7 +993,7 @@ class Map(ipyleaflet.Map):
             self.substitute_layer(old_basemap, basemap_tiles[basemap_name])
 
         dropdown.observe(on_click, "value")
-        basemap_control = WidgetControl(widget=dropdown, position="topright")
+        basemap_control = ipyleaflet.WidgetControl(widget=dropdown, position="topright")
         self.add_control(basemap_control)
 
     def add_legend(
@@ -1175,7 +1174,7 @@ class Map(ipyleaflet.Map):
                     "overflow": "scroll",
                 }
             )
-            legend_control = WidgetControl(
+            legend_control = ipyleaflet.WidgetControl(
                 widget=legend_output_widget, position=position
             )
             legend_widget = widgets.HTML(value=legend_text)
@@ -1249,7 +1248,7 @@ class Map(ipyleaflet.Map):
             else:
                 colormap = colormap.to_step(3)
 
-        colormap_ctrl = WidgetControl(
+        colormap_ctrl = ipyleaflet.WidgetControl(
             widget=output,
             position=position,
             transparent_bg=transparent_bg,
@@ -1302,7 +1301,7 @@ class Map(ipyleaflet.Map):
 
         output = widgets.Output()
 
-        colormap_ctrl = WidgetControl(
+        colormap_ctrl = ipyleaflet.WidgetControl(
             widget=output,
[…]
