+++ mkdocs.yml
@@ -46,7 +46,7 @@ nav:
     - Home: index.md
     - Installation: installation.md
     - Get Started: get-started.md
-    - Usage: usage.md
+    - Usage: usage.ipynb
     - Tutorials: tutorials.md
     - Contributing: contributing.md
     - FAQ: faq.md
