+++ mkdocs.yml
@@ -46,7 +46,7 @@ nav:
     - Home: index.md
     - Installation: installation.md
     - Get Started: get-started.md
-    - Usage: usage.ipynb
+    - Usage: notebooks/00_key_features.ipynb
     - Tutorials: tutorials.md
     - Contributing: contributing.md
     - FAQ: faq.md
@@ -64,6 +64,7 @@ nav:
           - osm module: osm.md
           - toolbar module: toolbar.md
     - Notebooks:
+          - notebooks/00_key_features.ipynb
           - notebooks/01_leafmap_intro.ipynb
           - notebooks/02_using_basemaps.ipynb
           - notebooks/03_cog_stac.ipynb
