### 2021-07-17 868546f
Improved add_ahp and add_kml for http
