+++ leafmap/leafmap.py
@@ -1609,9 +1609,25 @@ class Map(ipyleaflet.Map):
         Raises:
             FileNotFoundError: The provided shapefile could not be found.
         """
-        in_shp = os.path.abspath(in_shp)
-        if not os.path.exists(in_shp):
-            raise FileNotFoundError("The provided shapefile could not be found.")
+
+        import glob
+
+        if in_shp.startswith("http") and in_shp.endswith(".zip"):
+            out_dir = os.path.abspath("./cache/shp")
+            if not os.path.exists(out_dir):
+                os.makedirs(out_dir)
+            download_from_url(in_shp, out_dir=out_dir, verbose=False)
+            files = list(glob.glob(os.path.join(out_dir, "*.shp")))
+            if len(files) > 0:
+                in_shp = files[0]
+            else:
+                raise FileNotFoundError(
+                    "The downloaded zip file does not contain any shapefile in the root directory."
+                )
+        else:
+            in_shp = os.path.abspath(in_shp)
+            if not os.path.exists(in_shp):
+                raise FileNotFoundError("The provided shapefile could not be found.")
 
         geojson = shp_to_geojson(in_shp)
         self.add_geojson(
@@ -1986,9 +2002,19 @@ class Map(ipyleaflet.Map):
             FileNotFoundError: The provided KML file could not be found.
         """
 
-        in_kml = os.path.abspath(in_kml)
-        if not os.path.exists(in_kml):
-            raise FileNotFoundError("The provided KML file could not be found.")
+        if in_kml.startswith("http") and in_kml.endswith(".kml"):
+            out_dir = os.path.abspath("./cache")
+            if not os.path.exists(out_dir):
+                os.makedirs(out_dir)
+            download_from_url(in_kml, out_dir=out_dir, unzip=False, verbose=False)
+            in_kml = os.path.join(out_dir, os.path.basename(in_kml))
+            if not os.path.exists(in_kml):
+                raise FileNotFoundError("The downloaded kml file could not be found.")
+        else:
+            in_kml = os.path.abspath(in_kml)
+            if not os.path.exists(in_kml):
+                raise FileNotFoundError("The provided KML could not be found.")
+
         self.add_vector(
             in_kml,
             layer_name,
