### 2021-07-17 a14bdf8
Improved add_ahp and add_kml for http
