### 2021-07-18 639b891
Improved key features notebook #83
