### 2021-07-18 55fdd85
Improved key features notebook #83
