### 2021-07-18 32d2d56
Added plotting backend comparison #83
