### 2021-07-18 02a0b73
Added plotting backend comparison #83
