### 2021-07-18 92ba1cc
Added plotting backend comparison #83
