### 2021-07-19 92e4e7a
Improved WhiteboxTools
