### 2021-07-20 dc206cf
Added leafmap and foliumap unittest #83
