+++ leafmap/foliumap.py
@@ -1375,6 +1375,10 @@ class Map(folium.Map):
         Returns:
             str: A string containing the HTML code.
         """
+
+        if self.options["layersControl"]:
+            self.add_layer_control()
+
         if outfile is not None:
             if not outfile.endswith(".html"):
                 raise ValueError("The output file extension must be html.")
@@ -1416,37 +1420,71 @@ class Map(folium.Map):
             )
         self.get_root().html.add_child(folium.Element(title_html))
 
-    def add_colormap(self, **kwargs):
+    def add_colormap(
+        self,
+        cmap="gray",
+        colors=None,
+        discrete=False,
+        label=None,
+        width=8.0,
+        height=0.4,
+        orientation="horizontal",
+        vmin=0,
+        vmax=1.0,
+        axis_off=False,
+        show_name=False,
+        font_size=12,
+        transparent_bg=False,
+        position="bottomright",
+        **kwargs,
+    ):
         """Adds a matplotlib colormap to the map."""
         raise NotImplementedError(
             "The folium plotting backend does not support this function. Use the ipyleaflet plotting backend instead."
         )
 
-    def add_maker_cluster(self, **kwargs):
+    def add_marker_cluster(self, event="click", add_marker=True):
         """Captures user inputs and add markers to the map."""
         raise NotImplementedError(
             "The folium plotting backend does not support this function. Use the ipyleaflet plotting backend instead."
         )
 
-    def add_minimap(self, **kwargs):
+    def add_minimap(self, zoom=5, position="bottomright"):
         """Adds a minimap (overview) to the ipyleaflet map."""
         raise NotImplementedError(
             "The folium plotting backend does not support this function. Use the ipyleaflet plotting backend instead."
         )
 
-    def add_point_layer(self, **kwargs):
+    def add_point_layer(
+        self, filename, popup=None, layer_name="Marker Cluster", **kwargs
+    ):
         """Adds a point layer to the map with a popup attribute."""
         raise NotImplementedError(
             "The folium plotting backend does not support this function. Use the ipyleaflet plotting backend instead."
         )
 
-    def add_raster(self, **kwargs):
+    def add_raster(
+        self,
+        image,
+        bands=None,
+        layer_name=None,
+        colormap=None,
+        x_dim="x",
+        y_dim="y",
+    ):
         """Adds a local raster dataset to the map."""
         raise NotImplementedError(
             "The folium plotting backend does not support this function. Use the ipyleaflet plotting backend instead."
         )
 
-    def add_time_slider(self, **kwargs):
+    def add_time_slider(
+        self,
+        layers_dict={},
+        labels=None,
+        time_interval=1,
+        position="bottomright",
+        slider_length="150px",
+    ):
         """Adds a time slider to the map."""
         raise NotImplementedError(
             "The folium plotting backend does not support this function. Use the ipyleaflet plotting backend instead."
@@ -1454,6 +1492,9 @@ class Map(folium.Map):
 
     def add_vector_tile_layer(
         self,
+        url="https://tile.nextzen.org/tilezen/vector/v1/512/all/{z}/{x}/{y}.mvt?api_key=gCZXZglvRQa6sB2z7JzL1w",
+        attribution="",
+        vector_tile_layer_styles=dict(),
         **kwargs,
     ):
         """Adds a VectorTileLayer to the map."""
@@ -1463,50 +1504,84 @@ class Map(folium.Map):
 
     def add_xy_data(
         self,
-        **kwargs,
+        in_csv,
+        x="longitude",
+        y="latitude",
+        label=None,
+        layer_name="Marker cluster",
     ):
         """Adds points from a CSV file containing lat/lon information and display data on the map."""
         raise NotImplementedError(
[…]
