### 2021-07-20 c6dc1ee
Added leafmap and foliumap unittest #83
