### 2021-07-20 bbef39e
Added leafmap and foliumap unittest #83
