+++ tests/test_leafmap.py
@@ -1,13 +1,16 @@
 #!/usr/bin/env python
 
-"""Tests for `leafmap` package."""
+"""Tests for `leafmap` module."""
 
+import os
 import unittest
 from leafmap import leafmap
+import geopandas as gpd
+from unittest.mock import patch
 
 
 class TestLeafmap(unittest.TestCase):
-    """Tests for `leafmap` package."""
+    """Tests for `leafmap` module."""
 
     def setUp(self):
         """Set up test fixtures, if any."""
@@ -15,12 +18,462 @@ class TestLeafmap(unittest.TestCase):
     def tearDown(self):
         """Tear down test fixtures, if any."""
 
-    def test_basemap_tiles(self):
+    def test_add_basemap(self):
+        """Check basemaps"""
+        m = leafmap.Map()
+        m.add_basemap("TERRAIN")
+        out_str = m.to_html()
+        assert "Google Terrain" in out_str
 
-        self.assertIsInstance(leafmap.basemap_tiles.to_dict(), dict)
+    def test_add_cog_layer(self):
+        """Check COG layer"""
+        m = leafmap.Map()
+        url = "https://opendata.digitalglobe.com/events/california-fire-2020/pre-event/2018-02-16/pine-gulch-fire20/1030010076004E00.tif"
+        m.add_cog_layer(url, name="Fire (pre-event)")
+        out_str = m.to_html()
+        assert "Fire (pre-event)" in out_str
+
+    def test_add_cog_mosaic(self):
+        """Check COG mosaic"""
+        m = leafmap.Map()
+        links = [
+            "https://opendata.digitalglobe.com/events/california-fire-2020/pre-event/2018-02-16/pine-gulch-fire20/1030010076004E00.tif",
+            "https://opendata.digitalglobe.com/events/california-fire-2020/pre-event/2018-08-18/pine-gulch-fire20/1040010041D3B300.tif",
+        ]
+
+        m.add_cog_mosaic(links, name="COG mosaic", attribution="MAXAR")
+        out_str = m.to_html()
+        assert "COG mosaic" in out_str
+        assert "MAXAR" in out_str
+
+    def test_add_colorbar(self):
+        """Check colorbar"""
+        m = leafmap.Map()
+        colors = ["006633", "E5FFCC", "662A00", "D8D8D8", "F5F5F5"]
+        vmin = 0
+        vmax = 4000
+        m.add_colorbar(colors=colors, vmin=vmin, vmax=vmax, caption="Elevation")
+        out_str = m.to_html()
+        assert "Elevation" in out_str
+
+    @patch("matplotlib.pyplot.show")
+    def test_add_colormap(self, mock_show):
+        """Check colormap"""
+        m = leafmap.Map()
+        m.add_colormap(cmap="gray", label="Elevation")
+        out_str = m.to_html()
+        assert "Elevation" in out_str
+
+    def test_add_gdf(self):
+        """Check GeoDataFrame"""
+        m = leafmap.Map()
+        gdf = gpd.read_file(
+            "https://github.com/giswqs/leafmap/raw/master/examples/data/cable-geo.geojson"
+        )
+        m.add_gdf(gdf, layer_name="Cable lines")
+        out_str = m.to_html()
+        assert "Cable lines" in out_str
+
+    def test_add_gdf_from_postgis(self):
+        """Check PostGIS"""
+        m = leafmap.Map()
+        try:
+            con = leafmap.connect_postgis(
+                database="nyc",
+                host="localhost",
+                user=None,
+                password=None,
+                use_env_var=True,
+            )
+            sql = "SELECT * FROM nyc_neighborhoods"
+            gdf = leafmap.read_postgis(sql, con)
+            m.add_gdf(gdf, layer_name="NYC")
+            m.add_gdf_from_postgis(
+                sql,
[…]
