+++ requirements.txt
@@ -3,7 +3,7 @@ colour
 folium>=0.12.0
 geojson
 googledrivedownloader
-ipyevents
+ipyevents<=0.9.0
 ipyleaflet
 matplotlib
 mss
