+++ .github/workflows/build.yml
@@ -38,5 +38,8 @@ jobs:
                   pip install -r requirements.txt -r requirements_dev.txt
             - name: PKG-TEST
               run: |
-                  pip install geopandas osmnx
+                  pip install geopandas osmnx xarray_leaflet
                   python -m unittest discover tests/
+              env:
+                  HEREMAPS_API_KEY: ${{ secrets.HEREMAPS_API_KEY }}
+                  PLANET_API_KEY: ${{ secrets.PLANET_API_KEY }}
