+++ .github/workflows/docs.yml
@@ -21,11 +21,10 @@ jobs:
               run: codespell --skip="*.csv,*.geojson,*.json,*.js" --ignore-words-list="aci,acount,acounts,fallow,hart,hist,nd,ned,ois,wqs"
             - name: PKG-TEST
               run: |
-                  pip install geopandas osmnx
+                  pip install xarray_leaflet geopandas osmnx sqlalchemy psycopg2
                   python -m unittest discover tests/
             - run: python -m pip install --upgrade pip
-            - run: pip install mkdocs==1.1.2 mkdocs-material mkdocstrings mkdocs-git-revision-date-plugin mkdocs-jupyter ipykernel livereload
-            - run: pip install xarray_leaflet geopandas osmnx sqlalchemy psycopg2
+            - run: pip install mkdocs mkdocs-material mkdocstrings mkdocs-git-revision-date-plugin mkdocs-jupyter ipykernel livereload
             - run: mkdocs gh-deploy --force
               env:
                   USE_FOLIUM: ${{ secrets.USE_FOLIUM }}
