### 2021-07-21 64a1a6e
Improved JOSS paper #83
