+++ paper/paper.md
@@ -19,32 +19,46 @@ bibliography: paper.bib
 
 # Summary
 
-**Leafmap** is a Python package for interactive mapping and geospatial analysis with minimal coding in a Jupyter environment. It is a spin-off project of the [geemap](https://geemap.org) Python package [@Wu2020], which was designed specifically to work with [Google Earth Engine](https://earthengine.google.com) (GEE) [@Gorelick2017]. However, not everyone in the geospatial community has access to the GEE cloud computing platform. Leafmap is designed to fill this gap for non-GEE users. It is a free and open-source Python package that enables users to analyze and visualize geospatial data with minimal coding in a Jupyter environment, such as Google Colab, Jupyter notebook, and JupyterLab. Leafmap is built upon a number of open-source packages, such as [folium](https://github.com/python-visualization/folium) [@Filipe2021], [ipyleaflet](https://github.com/jupyter-widgets/ipyleaflet) [@Renou2021], and [here-map-widget-for-jupyter](https://github.com/heremaps/here-map-widget-for-jupyter) [@Kharude2021] (for creating interactive maps), [WhiteboxTools](https://github.com/jblindsay/whitebox-tools) [@Lindsay2018] and [whiteboxgui](https://github.com/giswqs/whiteboxgui) (for analyzing geospatial data), and [ipywidgets](https://github.com/jupyter-widgets/ipywidgets) [@Grout2021] (for designing interactive graphical user interface [GUI]). The WhiteboxTools library currently contains **470+** tools for advanced geospatial analysis, which are each grouped based on their main function into one of the following categories. Users can access these tools via whiteboxgui directly within the leafmap user interface without writing a single line of code.
-
--   [Data Tools](https://jblindsay.github.io/wbt_book/available_tools/data_tools.html)
--   [Geomorphometric Analysis](https://jblindsay.github.io/wbt_book/available_tools/geomorphometric_analysis.html)
--   [GIS Analysis](https://jblindsay.github.io/wbt_book/available_tools/gis_analysis.html)
--   [Hydrological Analysis](https://jblindsay.github.io/wbt_book/available_tools/hydrological_analysis.html)
--   [Image Processing Tools](https://jblindsay.github.io/wbt_book/available_tools/image_processing_tools.html)
--   [LiDAR Tools](https://jblindsay.github.io/wbt_book/available_tools/lidar_tools.html)
--   [Mathematical and Statistical Analysis](https://jblindsay.github.io/wbt_book/available_tools/mathand_stats_tools.html)
--   [Precision Agriculture](https://jblindsay.github.io/wbt_book/available_tools/precision_agriculture.html)
--   [Stream Network Analysis](https://jblindsay.github.io/wbt_book/available_tools/stream_network_analysis.html)
[…]
