### 2021-07-21 b453889
Improved JOSS paper #83
