+++ paper/paper.bib
@@ -104,3 +104,23 @@
   year        =  2021,
   url         = "https://github.com/heremaps/here-map-widget-for-jupyter"
 }
+
+
+@MISC{Mease2021,
+  title       = "{plotly.py: The interactive graphing library for Python}",
+  author      = "Mease, Jon and {et al.}",
+  institution = "Github",
+  year        =  2021,
+  url         = "https://github.com/plotly/plotly.py"
+}
+
+
+
+@MISC{He2021,
+  title       = "{kepler.gl: A powerful open source geospatial analysis tool for
+                 large-scale data sets}",
+  author      = "He, Shan and {et al.}",
+  institution = "Github",
+  year        =  2021,
+  url         = "https://github.com/keplergl/kepler.gl"
+}
