### 2021-07-21 8f18ab0
Improved JOSS paper #83
