### 2021-07-21 #3414 @giswqs
@martinfleis @TomasBeuzen Thank you again for your constructive comments. I have revised the repo, documentation and paper based on your comments. You can check out the commit history referenced in https://github.com/giswqs/leafmap/issues/83. Each commit addresses one of your comments, in no particular order. Please review it and let me know if you have any further concerns. Thank you very much for your time.
