+++ setup.py
@@ -53,6 +53,6 @@ setup(
     test_suite="tests",
     tests_require=test_requirements,
     url="https://github.com/giswqs/leafmap",
-    version="0.3.3",
+    version="0.3.4",
     zip_safe=False,
 )
