### 2021-07-24 661a908
Added unittest for toolbar module #83
