### 2021-07-24 88cdc13
Added unittest for toolbar module #83
