### 2021-07-24 afce12d
Added unittest for toolbar module #83
