+++ tests/test_leafmap.py
@@ -239,19 +239,19 @@ class TestLeafmap(unittest.TestCase):
         out_str = m.to_html()
         assert "NYC buildings" in out_str
 
-    # def test_add_planet_by_month(self):
-    #     """Check Planet monthly imagery"""
-    #     m = leafmap.Map()
-    #     m.add_planet_by_month(year=2020, month=8)
-    #     out_str = m.to_html()
-    #     assert "Planet_2020_08" in out_str
+    def test_add_planet_by_month(self):
+        """Check Planet monthly imagery"""
+        m = leafmap.Map()
+        m.add_planet_by_month(year=2020, month=8)
+        out_str = m.to_html()
+        assert "Planet_2020_08" in out_str
 
-    # def test_add_planet_by_quarter(self):
-    #     """Check Planet quarterly imagery"""
-    #     m = leafmap.Map()
-    #     m.add_planet_by_quarter(year=2019, quarter=2)
-    #     out_str = m.to_html()
-    #     assert "Planet_2019_q2" in out_str
+    def test_add_planet_by_quarter(self):
+        """Check Planet quarterly imagery"""
+        m = leafmap.Map()
+        m.add_planet_by_quarter(year=2019, quarter=2)
+        out_str = m.to_html()
+        assert "Planet_2019_q2" in out_str
 
     def test_add_point_layer(self):
         """Check adding point layer"""
@@ -299,13 +299,13 @@ class TestLeafmap(unittest.TestCase):
         out_str = m.to_html()
         assert "Google Satellite" in out_str
 
-    # def test_add_time_slider(self):
-    #     """Check adding time slider"""
-    #     m = leafmap.Map()
-    #     layers_dict = leafmap.planet_quarterly_tiles()
-    #     m.add_time_slider(layers_dict, time_interval=1)
-    #     out_str = m.to_html()
-    #     assert "Planet_2019_q2" in out_str
+    def test_add_time_slider(self):
+        """Check adding time slider"""
+        m = leafmap.Map()
+        layers_dict = leafmap.planet_quarterly_tiles()
+        m.add_time_slider(layers_dict, time_interval=1)
+        out_str = m.to_html()
+        assert "Planet_2019_q2" in out_str
 
     def test_add_vector(self):
         """Check adding vector"""
