### 2021-07-24 #3414 @martinfleis
@giswqs Thanks a lot for the changes, great job! Though I could imagine higher code coverage I don't think it an issue blocking the publication. All the other comments have been resolved.

I have ticked remaining boxes in the review form above and recommend the paper for publication.
