### 2021-07-25 #3414 @giswqs
@martinfleis Thank you very much for taking the time to re-review the package. I agree with you that the code coverage can be improved. I tried my best. There are some interactive custom widgets that are a bit challenging for unit tests. I will continue to improve the code coverage in the future when unit tests for these interactive custom widgets become possible.

@TomasBeuzen Please let me know if there is anything else you would like me to address. Thank you.
