### 2021-07-25 #3414 @TomasBeuzen
@giswqs (cc @ajstewartlang) - you've (comprehensive!) addressed all my comments, and thank you for laying everything out in easy-to-navigate commits. I'm satisfied that the submissions meets all the JOSS criteria and have ticked all my boxes. Excellent work again @giswqs for this excellent submission - I look forward to following how the software develops and is used by the community into the future.
