+++ .github/workflows/docs.yml
@@ -21,7 +21,7 @@ jobs:
               run: codespell --skip="*.csv,*.geojson,*.json,*.js" --ignore-words-list="aci,acount,acounts,fallow,hart,hist,nd,ned,ois,wqs"
             - name: PKG-TEST
               run: |
-                  pip install xarray_leaflet geopandas osmnx sqlalchemy psycopg2
+                  pip install xarray_leaflet geopandas osmnx sqlalchemy psycopg2 keplergl
                   python -m unittest discover tests/
               env:
                   PLANET_API_KEY: ${{ secrets.PLANET_API_KEY }}
