+++ .github/workflows/build.yml
@@ -38,7 +38,7 @@ jobs:
                   pip install -r requirements.txt -r requirements_dev.txt
             - name: PKG-TEST
               run: |
-                  pip install geopandas osmnx xarray_leaflet
+                  pip install geopandas osmnx xarray_leaflet keplergl
                   python -m unittest discover tests/
               env:
                   HEREMAPS_API_KEY: ${{ secrets.HEREMAPS_API_KEY }}
