+++ mkdocs.yml
@@ -90,3 +90,4 @@ nav:
           - notebooks/23_colormaps.ipynb
           - notebooks/24_heatmap.ipynb
           - notebooks/25_map_title.ipynb
+          - notebooks/26_kepler_gl.ipynb
