### 2021-07-26 b1e6011
Update paper.md
spotted an agreement error - 'plethora' is a singular noun
