+++ setup.py
@@ -53,6 +53,6 @@ setup(
     test_suite="tests",
     tests_require=test_requirements,
     url="https://github.com/giswqs/leafmap",
-    version="0.3.4",
+    version="0.3.5",
     zip_safe=False,
 )
