### 2021-07-26 #3414 @ajstewartlang
> @whedon generate pdf

I've reached out to the rest of the JOSS team to find out why `whedon` isn't generating the pdf...
