### 2021-07-26 #3414 @giswqs
@ajstewartlang Thank you for fixing the typo. I have made a tag version that includes all changes made to the paper.

- Tagged version: v0.3.5
- Zenodo link: https://doi.org/10.5281/zenodo.5137372
- DOI: 10.5281/zenodo.5137372
