### 2021-07-26 #3414 @ajstewartlang
Many thanks @martinfleis and @TomasBeuzen for your helpful and detailed reviews - thank you @giswqs for addressing all the comments - I've read through myself and agree that this is a great submission that will be very useful to many people! Congrats and thank you for submitting to JOSS!
